Only check repetition nesting (#2)
I initially thought that the transcriber repetition stack must be equal
to the matcher repetition stack, but it turns out that only the nesting
value must be equal. In other words, this means that the following is a
valid macro:

```rust
macro_rules! test {
    // Matched repetition stack: +?*
    ( $( $( $( $a:ident, )* )? )+ ) => {
        // Transcribed repetition stack: ***
        $( $( $( $a, )* )* )*
    }
}
```

In addition, the error message has been reworked. It's probably a bit
too long, but I don't see how we can improve that.

---------

Signed-off-by: Sasha Pourcelot <sasha.pourcelot@protonmail.com>