Add a MSRV policy (#3)
It turns out that the minimal version that allows `expandable` to
compile is `1.65.0` (it introduced `let-else` and GATs), so let's make
it our MSRV.

A few key points:
- The crate is guaranteed to compile for Rust 1.65, but it will support
syntax added after Rust 1.65 as well. This allows us to support new
syntax as patch versions,
- The error messages are unstable (the wording may change at any moment,
without notice). This allows us to rework them as much as possible,
- Adding new possible errors is a breaking change.

Many thanks to the [`cargo-msrv`] folks. This tool rules!

[`cargo-msrv`]: https://github.com/foresterre/cargo-msrv

---------

Signed-off-by: Sasha Pourcelot <sasha.pourcelot@protonmail.com>