Add debug-only span (#6)
The test suite previously used `()` as span. This makes debugging a
pain, as the only information we have about a token is `()`. This commit
fixes this by using a custom debug-only `DebugSpan` instead. This type
wraps an `usize` that starts at `0` at each call to `quote` and is
monotonically increasing:

```rs
quote! {
    fn /* 0 */ h2g2 /* 1 */ ( /* 2 */ ) /* 2 */ -> /* 3 */ usize /* 4 */ { /* 5 */
        42 /* 6 */
    } /* 5 */
}
```

This commit will make development _massively_ easier.

---------

Signed-off-by: Sasha Pourcelot <sasha.pourcelot@protonmail.com>