fix: Add the missing AfterIf transitions (#17)
(This is just another instance of me not being able to update multiple
things at the same time)

#14 added a bunch of transitions that occurred in the `AfterExpr` state.
I forgot to add them in the `AfterIf` state as well. This PR fixes that.

Additionally I added a test so that the CI will break if I ever forget
to update the `AfterIf` state. Great!

---------

Signed-off-by: Sasha Pourcelot <sasha.pourcelot@protonmail.com>