refactor: add a transition inheritance system (#21)
Many state transitions set are a superset of another state's transition
set. The current solution for this is to write these transitions one
time for each state. This is very suboptimal and error-prone (see #17).

This PR fixes the situation by allowing a state to "extend" another set,
and inherit its states. This allows us to remove almost all the
transitions in the `AfterIf` state, and will make adding new supersets
of `ExprStart` easier.

Signed-off-by: Sasha Pourcelot <sasha.pourcelot@protonmail.com>