feat: converge on transitions instead of converging on final state (#34)
The parser generated by the `gramar_gen` crate works well for parsing,
but the generated state machine pushes _a lot_ of symbols on its stack.
This is fine when parsing, because these symbols are consumed later in
the parsing. This is less ok when we compare multiple repetitions of the
same sequence of tokens, because each repetition is likely to add more
symbols in the stack.

In order to mitigate this, the state machine matcher converges on
transitions (ie: how a state has changed when fed with tokens) rather
than on the final state itself. That's the purpose of this MR. This
relies on the fact that applying a stream of macro on a state will
result on the same transition, no matter how much time the stream has
been applied before.

This approach is much closer to mathematical induction: if we see that
applying one time a repetition yields a valid state ( $P(0)$ ) and
manage to prove that for any amount of repetition, adding one more
repetition yields the same transition ( $P(n) \implies P(n + 1)$ ), then
the resulted expansion is valid for any number of repetition.

Will help for #29.

---------

Signed-off-by: Sasha Pourcelot <sasha.pourcelot@protonmail.com>