feat: avoid printing a comically large amount of tokens on error (#41)
The previous error reporting mechanism was just "print all expected
tokens, separated by commas". This led to situations where a comically
high amount of tokens were written in the error message. For instance:

```
error: Potentially invalid expansion. Expected `!=`, `%`, `&&`, `&`, `*`, `+`, `,`, `-`, `..=`, `..`, `.`, `/`, `<<`, `<=`, `<`, `==`, `>=`, `>>`, `>`, `^`, `|`, `||`, a `(`, a `)`.
  --> tests/ui/fail/invalid_fn_calls.rs:22:25
   |
22 |         $fn_name($arg1 $arg2)
   |                         ^^^^
```

This PR limits the amount of _printed_ tokens to 6 and adds a potential
"or _n_ other_s_" if needed. This changes the error message to:

```
error: Potentially invalid expansion. Expected `!=`, `%`, `&&`, `&`, `*`, `+` or 18 others.
  --> tests/ui/fail/invalid_fn_calls.rs:22:25
   |
22 |         $fn_name($arg1 $arg2)
   |                         ^^^^
```

Which is way clearer.

The threshold value may need to be tweaked.

---------

Signed-off-by: Sasha Pourcelot <sasha.pourcelot@protonmail.com>