refactor: use opaque type to represent state in transitions (#48)
In `rust-grammar-dpdfa`, we represent the states pushed during a
transition as `&'static str`. This is not perfect, because 1.
comparing/hashing two strings ain't cheap and 2. we leak the state names
to the library user.

This has been fixed by using a custom opaque type that contains a raw
pointer corresponding to the state itself.

---------

Signed-off-by: Sasha Pourcelot <sasha.pourcelot@protonmail.com>