refactor: make the expansion check fully deterministic (#50)
We did not care about `expandable`'s deterministicness earlier because
there our focus was on "is there an error and where". With #49, we also
need to solve the question "how did we get here", which requires having
a fixed order when checking a macro.

(Also I'm almost sure our testing is not optimal - otherwise we may have
discovered this earlier)

Implementation-wise, this was fixed by switching from a `HashSet` to a
`BTreeSet`. I believe this won't cause any performance issue because we
will still have a low amount of states, resulting in a low amount of
allocations anywhere. We are probably loosing big-O-wise, but we don't
care about big-O because our numbers are smol enough.

---------

Signed-off-by: Sasha Pourcelot <sasha.pourcelot@protonmail.com>