feat: match expressions (#58)
Title says everything.

Sadly the support is not as complete as I'd like it to be.

It turns out it's incredibly hard to guess if the operand is a block
expression or not. This requires many refactorings that I'm too tired to
do. The result is that the `,` is required at the end of the `match`
arm, even if the operand is a block expression.

Part of #29.

---------

Signed-off-by: Sasha Pourcelot <sasha.pourcelot@protonmail.com>