refactor: add variable support in `rust-grammar-dpdfa`/`grammar-gen` (#61)
Well, it turns out it is not that convenient to write a Rust parser
without using variables. Who could predict? :clown_face:

This MR completely changes `grammar-gen` and `rust-grammar-dpdfa` so
that functions can take arguments. Internally, `rust-grammar-dpdfa` is
now backed by a virtual machine that runs its own bytecode, and
`grammar-gen` is a compiler that targets this very bytecode. I may add
loops in the future, I don't know.

For what it's worth, the previous approach was to use a pushdown
automation, which did not allow for variables and return values.

I'm actually surprised that most of the debugging I had to do was in the
virtual machine implementation, and I got the codegen correct on the
first attempt.

This is going to make my work on #29 massively simpler.