feat: allow variables in more places, add support for && (#63)
Previous iteration of `grammar-gen` forced the function arguments to be
atoms. This MR adds support for variables as well, allowing the
following code to be compiled:

```
let a = Atom;
fn_call(a);
```

This MR also adds support for the `&&` and `||` operators. I had to do a
bit of codegen refactoring, which resulted in less instructions being
emitted when a condition is evaluated.