feat: add minimal support for the Lifetime token description (#72)
For some reason, I completely forgot to add the Lifetime token
description, which disabled me from writing parts of the parser.

This MR does the minimal amount of work on this side, and adds it in
`grammar-gen` and in `rust-grammar-dpdfa`. I don't think the lifetime is
correctly parsed in the frontend. This will be done later, in another PR
:sparkles:.